refactor: 높낮이 계산 로직 분리
